Adds Mask to Distance Field, moves all nodes into their own directory
