+++ ui/client/dagpipes/overview.css
@@ -6,7 +6,7 @@
 .react-flow__node-save, .react-flow__node-load,
 .react-flow__node-sum, .react-flow__node-multiply,
 .react-flow__node-filter_by_country, .react-flow__node-threshold,
-.react-flow__node-scalar_operation
+.react-flow__node-scalar_operation, .react-flow__node-mask_distance
 {
   font-size: 10px;
   width: 180px;
@@ -24,7 +24,8 @@
 .react-flow__node-load,
 .react-flow__node-save,
 .react-flow__node-threshold,
-.react-flow__node-scalar_operation {
+.react-flow__node-scalar_operation,
+.react-flow__node-mask_distance {
   width: 16rem;
 }
